requirements.txt
streamlit
groq
python-dotenv
yfinance
lxml
html5lib
beautifulsoup4
tabulate